+++ R/RcppExports.R
@@ -89,8 +89,8 @@ C_lasrange <- function(las, flightlines) {
     .Call(`_lidR_C_lasrange`, las, flightlines)
 }
 
-C_local_maximum <- function(las, ws, ncpu) {
-    .Call(`_lidR_C_local_maximum`, las, ws, ncpu)
+C_local_maximum <- function(las, ws, filter, ncpu) {
+    .Call(`_lidR_C_local_maximum`, las, ws, filter, ncpu)
 }
 
 C_isolated_voxel <- function(las, res, isolated) {